Add final tests

+++ bookkeeper/models/budget.py
@@ -19,30 +19,30 @@ class Budget:
     period- срок
     pk - id записи в базе данных
     """
-    amount: int
+    amount: float
     period: str
     category: int | None = None
-    restrict: int | None = None
+    restrict: float | None = None
     pk: int = 0
 
     @classmethod
     def get_amounts(cls,
-                    repo: AbstractRepository['Expense']) -> list[int]:
+                    repo: AbstractRepository['Expense']) -> list[float]:
         "Функция для подсчёта суммы расходов за день, неделю, месяц"
-        day_amount = 0
+        day_amount = 0.0
         date = datetime.now()
         day_expenses = repo.get_all(where={"expense_date":
                                            date.strftime('%Y-%m-%d')})
         day_amount += sum(exp.amount for exp in day_expenses)
 
-        week_amount = 0
+        week_amount = 0.0
         for i in range(date.weekday() + 1):
             weekday = date - timedelta(days=i)
             weekday_expenses = repo.get_all(where={"expense_date":
                                                    weekday.strftime('%Y-%m-%d')})
             week_amount += sum(exp.amount for exp in weekday_expenses)
 
-        month_amount = 0
+        month_amount = 0.0
         for i in range(date.day):
             monthday = date - timedelta(days=i)
             month_expenses = repo.get_all(where={"expense_date":
@@ -53,7 +53,7 @@ class Budget:
     @classmethod
     def create_from_restricts(
             cls,
-            restricts: list[tuple[int | None, int | None]],
+            restricts: list[tuple[int | None, float | None]],
             repo: AbstractRepository['Expense']) -> list['Budget']:
         "Функция для создания списка бюджета из списка ограничений"
         periods = ('день', 'неделя', 'месяц')
